+++ milkman/src/main/java/milkman/ui/plugin/ContentTypePlugin.java
@@ -1,10 +1,9 @@
 package milkman.ui.plugin;
 
-import java.util.Collection;
-
+import milkman.ui.components.CodeFoldingContentEditor.ContentRange;
 import org.fxmisc.richtext.model.StyleSpans;
 
-import milkman.ui.components.CodeFoldingContentEditor.ContentRange;
+import java.util.Collection;
 
 /**
 * extension point for content types used by the content-editor
@@ -38,8 +37,8 @@ public interface ContentTypePlugin {
     * see milkman-syntax.css for style-classes that you can use in your highlighting.
     */
 	StyleSpans<Collection<String>> computeHighlighting(String text);
-	
-	
+
+
 	/**
 	 * returns true if code folding is supported
 	 */
@@ -54,5 +53,10 @@ public interface ContentTypePlugin {
 	 */
 	default ContentRange computeFolding(String text) {
 		throw new UnsupportedOperationException();
+	}
+
+
+	default String computeIndentationForNextLine(String currentLine){
+		return "";
 	}
 }